firmware: rewrite delayMicroseconds to support longer delays
Now it supports delays up to 5 minutes, though it is definitely not
a good idea to use it like that.

Also previous version had a problem when delay is close to USHRT_MAX,
it could run for several more timer periods before break.

+++ firmware/src/timer.h
@@ -10,6 +10,6 @@ uint32_t sysClockFreq;
 uint32_t ticksPerMicrosecond;
 
 void initMicrosecondsDelayTimer(void);
-void delayMicroseconds(uint16_t microseconds);
+void delayMicroseconds(uint32_t microseconds);
 
 void initDebugTimer(void);
